Move GC stats helpers to source file
