+++ PCbuild/_remote_debugging.vcxproj.filters
@@ -15,6 +15,9 @@
     <ClCompile Include="..\Modules\_remote_debugging\module.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\Modules\_remote_debugging\gc_stats.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="..\Modules\_remote_debugging\object_reading.c">
       <Filter>Source Files</Filter>
     </ClCompile>
